# Switch from `haskell-src-exts` to `ghc-lib-parser`