--- a/.hindent.yaml
+++ b/.hindent.yaml
@@ -1,3 +1,6 @@
 tab-size: 2
 line-length: 80
 force-trailing-newline: true
+extensions:
+  - TypeApplications
+  - PatternSynonyms