--- a/src/HIndent/Pretty/Combinators/Comment.hs
+++ b/src/HIndent/Pretty/Combinators/Comment.hs
@@ -1,0 +1,12 @@
+-- | Printer combinators for handling comments.
+module HIndent.Pretty.Combinators.Comment
+  ( eolCommentsArePrinted
+  ) where
+
+import           Control.Monad.State
+import           HIndent.Types
+
+-- | Claims that comments were printed. Next time calling 'string' will
+-- print a newline before printing a text.
+eolCommentsArePrinted :: Printer ()
+eolCommentsArePrinted = modify (\s -> s {psEolComment = True})