--- a/src/main/Benchmark.hs
+++ b/src/main/Benchmark.hs
@@ -13,6 +13,7 @@ import qualified Data.ByteString.Builder as S
 import qualified Data.ByteString.Char8 as S8
 import qualified Data.ByteString.UTF8 as UTF8
 import           HIndent
+import           HIndent.LanguageExtension
 import           HIndent.Types
 import           Markdone