… TESTS.md
… src/HIndent.hs
… src/HIndent/ModulePreprocessing.hs
… src/HIndent/ModulePreprocessing/CommentRelocation.hs
… src/HIndent/Pragma.hs
… src/HIndent/Pretty.hs
… src/HIndent/Pretty/Combinators.hs
… src/HIndent/Pretty/Combinators/Getter.hs
… src/HIndent/Pretty/Combinators/Indent.hs
… src/HIndent/Pretty/Combinators/Lineup.hs
… src/HIndent/Pretty/Combinators/Op.hs
… src/HIndent/Pretty/Combinators/Outputable.hs
… src/HIndent/Pretty/Combinators/RhsSeparator.hs
… src/HIndent/Pretty/Combinators/String.hs
… src/HIndent/Pretty/Combinators/Switch.hs
… src/HIndent/Pretty/Combinators/Wrap.hs
… src/HIndent/Pretty/Import.hs
… src/HIndent/Pretty/Import/Sort.hs
… src/HIndent/Pretty/NodeComments.hs
… src/HIndent/Pretty/Pragma.hs
… src/HIndent/Pretty/SigBindFamily.hs
… src/HIndent/Pretty/Types.hs
… src/HIndent/Types.hs
… src/main/Main.hs
… src/main/Markdone.hs
… src/main/Path/Find.hs
… src/main/Test.hs
… stack.yaml
… stack.yaml.lock