See #587 for the background.

### Problems
- Cannot handle `LINE` pragmas (e.g., `{-# LINE 1 "foo/bar.hs" #-}`). It causes a parse error.
- Cannot preserve comments around a pragma.
```haskell
{-# LANGUAGE CPP #-} -- Comments

module Foo where
```
- `stack bench` indicates that the HIndent after rewriting is much slower than the original one.
  - Currently, the source files are lexically parsed once to extract the extensions specified in the source code, and then parsed for the actual formatting. I believe this is one of the reasons for this.
- Some formatting depends on `ghc-lib-parser`'s internal formatted, and we should avoid it if possible. See the comment in `HIndent.Pretty.Combinators.Outputable`.

### Notes
Some formatting results differ from the one by the original HIndent due to several reasons.
- A bug in `haskell-src-exts` wrongly parses a left-associative infix operator as right-associative and vice versa. 
- The original HIndent used `haskell-src-exts`' internal formatter and reproducing it is difficult.

### Benchmark results

#### Before the switch

```
Build profile: -w ghc-9.2.5 -O1
In order, the following will be built (use -v for more details):
 - hindent-5.3.5 (bench:hindent-bench) (first run)
Preprocessing benchmark 'hindent-bench' for hindent-5.3.5..
Building benchmark 'hindent-bench' for hindent-5.3.5..
Running 1 benchmarks...
Benchmark hindent-bench: RUNNING...
benchmarking Large inputs/Bunch of declarations
time                 27.41 ms   (27.12 ms .. 27.72 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 27.09 ms   (26.61 ms .. 27.27 ms)
std dev              602.4 μs   (234.1 μs .. 1.116 ms)

benchmarking Large inputs/Bunch of declarations - sans comments
time                 2.817 ms   (2.809 ms .. 2.823 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.775 ms   (2.767 ms .. 2.783 ms)
std dev              25.82 μs   (22.48 μs .. 30.44 μs)

benchmarking Complex inputs/Quasi-quotes with nested lets and operators
time                 1.579 ms   (1.576 ms .. 1.582 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.590 ms   (1.586 ms .. 1.594 ms)
std dev              12.46 μs   (10.30 μs .. 15.51 μs)

benchmarking Complex inputs/Lots of comments and operators
time                 5.861 ms   (5.833 ms .. 5.881 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 5.839 ms   (5.820 ms .. 5.856 ms)
std dev              54.77 μs   (45.75 μs .. 66.22 μs)

Benchmark hindent-bench: FINISH
```

#### After the switch

```
Build profile: -w ghc-9.2.5 -O1
In order, the following will be built (use -v for more details):
 - hindent-5.3.5 (bench:hindent-bench) (first run)
Preprocessing benchmark 'hindent-bench' for hindent-5.3.5..
Building benchmark 'hindent-bench' for hindent-5.3.5..
Running 1 benchmarks...
Benchmark hindent-bench: RUNNING...
benchmarking Large inputs/Bunch of declarations
time                 158.1 ms   (157.0 ms .. 159.2 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 159.3 ms   (158.3 ms .. 162.4 ms)
std dev              2.152 ms   (355.4 μs .. 3.220 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking Large inputs/Bunch of declarations - sans comments
time                 157.2 ms   (153.6 ms .. 159.7 ms)
                     1.000 R²   (0.998 R² .. 1.000 R²)
mean                 157.2 ms   (155.1 ms .. 158.1 ms)
std dev              1.869 ms   (719.5 μs .. 3.124 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking Complex inputs/Quasi-quotes with nested lets and operators
time                 68.53 ms   (68.35 ms .. 68.78 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 68.49 ms   (68.29 ms .. 68.79 ms)
std dev              431.2 μs   (245.1 μs .. 637.3 μs)

benchmarking Complex inputs/Lots of comments and operators
time                 40.77 ms   (40.46 ms .. 41.18 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 40.64 ms   (40.45 ms .. 40.82 ms)
std dev              385.9 μs   (255.7 μs .. 612.2 μs)

Benchmark hindent-bench: FINISH
```

Fixes: #587 
Fixes: #600 
Closes: #582  
Closes: #599