@toku-sa-n 2022-09-21
@mihaimaruseac Hello. I have a question. Do we still need to support GHC versions that are older than 8.10?

Though it's not impossible, supporting these versions is hard because `ghc-lib-parser` changed how it generates an AST from a given source. `ghc-lib-parser` of the compatible version with GHC older than 8.10 generates an AST having the old structure, and supporting old and new ones requires adding many codes, especially for handling comments.

I think supporting GHC 8.10.7 and 9.2.2 is enough. 8.10.7 is the current [recommended version](https://www.haskell.org/ghcup/install/#which-versions-get-installed) by `ghcup`, and 9.2.2 is supported currently and is newer than the recommended version.

Do you have any thoughts on this?