@mihaimaruseac 2022-09-21
Sorry, was not following the progress here too closely.

I think it's good to drop the old versions, I'll send a fix to remove them from CI right now