@toku-sa-n 2022-09-22
@mihaimaruseac It seems that building with GHC 9.4.1 fails because of [a GHC bug](https://gitlab.haskell.org/ghc/ghc/-/issues/21990). I've created #597 to remove the support of the version.