@mihaimaruseac 2022-09-22
Thank you for the other PR. From reading the issue (and the CI on #596, I shouldn't have merged it that early) it looks like only Windows is affected? Made a suggestion on your PR for that end.