@toku-sa-n 2022-09-22
Thanks. Yes, it only affects Windows. I've fixed it.