@toku-sa-n 2022-11-06
@mihaimaruseac There are still a few things to do (e.g., removing files used in the development from commits, exporting necessary symbols, etc.), and a few FIXMEs which I can't solve immediately, but I think the work of switching from `haskell-src-exts` to `ghc-lib-parser` is done.

I apologize that the diffs have become very large. Some of the diffs are not needed for this PR and can be submitted as separate PRs.

I have two thoughts.

* I think we need to clean up the `TESTS.md` file. In particular, I think it needs to be broken up into smaller sections. Also, the regression test section is huge and not classified at all. This section should be eliminated, and each test suite should be moved to the appropriate section. BTW I've submitted #606. Could you take a look at it?
* Do we need to re-export the functions for the tests exported in the `HIndent` module and the `parseMode`? I think `testAst` is sufficient for the former, and I don't see much need to export the latter.

Thank you.