@toku-sa-n 2022-12-23
Yes, it's ready for review! I was planning to mark this PR as ready after
running benchmarks, but you can start reviewing now.

On Sat, Dec 24, 2022, 01:31 Mihai Maruseac ***@***.***> wrote:

> Hi @toku-sa-n <https://github.com/toku-sa-n>.
>
> I see you merged all commits and removed the TODO-checklist. Does this
> mean this is ready for review? :)
>
> Sorry if I'm too eager and this is not yet ready :)
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/mihaimaruseac/hindent/pull/593#issuecomment-1364100321>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AKDAAUYZJXWDM3C6ZDI7WKDWOXHVHANCNFSM6AAAAAAQGIPOCE>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>